Google Gemini projects/Gemini-AI-Chatbot/requirements.txt
streamlit==1.29.0
google-generativeai==0.6.1
python-dotenv==1.0.1
ipython==8.17.0
